+++ packages/mpv-player/src/index.ts
@@ -1,3 +1,8 @@
 export { MpvPlayerView } from "./MpvPlayerView";
-export { mpvAudio, type MpvAudioProgress } from "./MpvAudio";
+export {
+  mpvAudio,
+  type MpvAudioProgress,
+  type MpvAudioError,
+  type MpvAudioBuffering,
+} from "./MpvAudio";
 export * from "./MpvPlayer.types";